backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
psycopg2-binary==2.9.9

# RAG 관련 패키지
qdrant-client==1.7.0
sentence-transformers==2.3.1
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
